% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_range.R
\name{isRustAvailable}
\alias{isRustAvailable}
\title{Checks for the availability of the Rust function}
\usage{
isRustAvailable()
}
\value{
Boolean depending on the presence of the Rust function
}
\description{
Checks for the availability of the Rust function
}
